add switch support
To make it work I had to create a new CaseList class that inherents from Node.

also, added a file t05.in with a simple test case

while at it, I found a bug when we pass arguments with
numbers or strings directly and not by variable that we should fix.